CREATED - Função encoded

+++ EuclidesAlgorithm.h
@@ -1,7 +1,7 @@
 #ifndef EuclidesAlgorithm
 #define EuclidesAlgorithm
 
-unsigned int euclidesDefault(unsigned int n1, unsigned int n2) {
+unsigned int euclidesDefault(unsigned long long n1, unsigned long long n2) {
     unsigned int res = n1 % n2;
     
     if(res == 1) return res;// Poupa +1 chamada desnecessária
@@ -10,15 +10,15 @@ unsigned int euclidesDefault(unsigned int n1, unsigned int n2) {
     return euclidesDefault(n2, res);
 }
 
-unsigned int euclidesExtended(unsigned int n1, unsigned int n2, unsigned int &s, unsigned int &t) {
+unsigned int euclidesExtended(unsigned long long n1, unsigned long long n2, unsigned long long &s, unsigned long long &t) {
     s = 1;
     t = 0;
 
-    unsigned int newS = 0, newT = 1;
-    unsigned int aux, q;
+    unsigned long long newS = 0, newT = 1;
+    unsigned long long aux, q;
 
     do {
-        q = (int) n1 / n2;
+        q = (unsigned int) n1 / n2;
 
         aux = s-(newS * q);
         s = newS;
